requirements.txt
python-dotenv==0.19.0
gradio==2.3.1
google-generativeai==0.1.0